@@ -4,13 +4,12 @@ version: 0.4.2
 homepage: https://github.com/GaspardMerten/flutter_auto_form
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
+  sdk: '>=2.15.0 <3.0.0'
   flutter: ">=1.17.0"
 
 dependencies:
-  smarter_text_field: ^0.0.4
-  dropdown_search: ^1.0.4
-  file_picker: ^4.1.4
+  smarter_text_field: ^0.0.6
+  dropdown_search: ^2.0.1
   flutter:
     sdk: flutter
 
