@@ -1,28 +1,39 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_auto_form/flutter_auto_form.dart';
+import 'package:flutter_auto_form/src/models/field/field_context.dart';
+import 'package:flutter_auto_form/src/widgets/fields/interface.dart';
 
-class SelectFieldWidget<T> extends StatelessWidget {
+class SelectFieldWidget<T extends Object> extends FieldStatefulWidget {
   const SelectFieldWidget({
     Key? key,
-    required this.value,
-    required this.textBuilder,
-    required this.values,
-    required this.onChanged,
+    required this.fieldContext,
   }) : super(key: key);
 
-  final T value;
-  final List<T> values;
-  final ValueChanged<T?> onChanged;
-  final String Function(T value) textBuilder;
+  @override
+  State<SelectFieldWidget<T>> createState() => _SelectFieldWidgetState<T>();
+
+  @override
+  final FieldContext fieldContext;
+}
+
+class _SelectFieldWidgetState<T extends Object>
+    extends State<SelectFieldWidget<T>> {
+  late final AFSelectField<T> field =
+      widget.fieldContext.field as AFSelectField<T>;
 
   @override
   Widget build(BuildContext context) {
     return DropdownButtonFormField<T>(
-      value: value,
-      onChanged: onChanged,
+      value: field.value,
+      onChanged: (e) {
+        setState(() {
+          widget.fieldContext.onChanged(e);
+        });
+      },
       items: [
-        for (final value in values)
+        for (final value in field.values)
           DropdownMenuItem(
-            child: Text(textBuilder(value)),
+            child: Text(field.textBuilder(value)),
             value: value,
           )
       ],
